refactor(web): split COE data layer for Next.js patterns
Separate raw API fetching from computed data with React cache to improve performance and follow Next.js 16 ORM best practices
